+++ __test__/units/auth.controller.spec.ts
@@ -3,16 +3,13 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { Test, TestingModule } from '@nestjs/testing';
 import * as dotenv from 'dotenv';
-import {
-  returnCreateUser,
-  returnLoginSuccess,
-} from '../../test/mocks/returnUserMock';
-import { USER_REPOSITORY } from '../core/constants/constants';
-import { userProviders } from '../users/user.providers';
-import { UsersModule } from '../users/users.module';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { JwtStrategy } from './jwt.strategy';
+import { AuthController } from '../../src/auth/auth.controller';
+import { AuthService } from '../../src/auth/auth.service';
+import { JwtStrategy } from '../../src/auth/jwt.strategy';
+import { USER_REPOSITORY } from '../../src/core/constants/constants';
+import { userProviders } from '../../src/users/user.providers';
+import { UsersModule } from '../../src/users/users.module';
+import { returnCreateUser, returnLoginSuccess } from '../mocks/returnUserMock';
 
 dotenv.config({
   path: process.env.NODE_ENV === 'test' ? '.env.test' : '.env',
